+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ ChangeLog
 - Only try to update the file if something changed
 - Add ``--check`` option to check if all the files are up-to-date with their generators
 - Improve diff shown with ``--verbose`` option
+- Use different exit codes for unsynced files and missing configuration
 
 
 0.1 (2019-11-07)